feat(parser): report an error when tag to limit file inclusion could not be found
compares the tags that were actually processed during the file inclusion
with the tags provided in the macro

also, move the file inclusion funcs in a separate file

fixes #393
