+++ testsupport/console.go
@@ -11,9 +11,10 @@ import (
 )
 
 // VerifyConsoleOutput verifies that the readable 'console' contains the exact given msg
-func VerifyConsoleOutput(console Readable, msg string) {
+func VerifyConsoleOutput(console Readable, msg string) { // TODO also add a param for the log level associated with the msg
 	GinkgoT().Logf(console.String())
 	out := make(map[string]interface{})
+	// TODO: wrap the content of 'console` in a JSON document to avoid parsing error if there are multiple lines
 	err := json.Unmarshal(console.Bytes(), &out)
 	Expect(err).ShouldNot(HaveOccurred())
 	Expect(out["level"]).Should(Equal("error"))
@@ -25,8 +26,9 @@ func VerifyConsoleOutput(console Readable, msg string) {
 // end of the test.
 func ConfigureLogger() (Readable, func()) {
 	fmtr := log.StandardLogger().Formatter
-
+	level := log.StandardLogger().Level
 	buf := bytes.NewBuffer(nil)
+	log.SetLevel(log.WarnLevel)
 	log.SetOutput(buf)
 	log.SetFormatter(&log.JSONFormatter{
 		DisableTimestamp: true,
@@ -34,6 +36,7 @@ func ConfigureLogger() (Readable, func()) {
 	return buf, func() {
 		log.SetOutput(os.Stdout)
 		log.SetFormatter(fmtr)
+		log.SetLevel(level)
 	}
 }
